do a lot of works :D

+++ backend/lib/db.py
@@ -1,15 +1,16 @@
-import os
 import re
 from typing import List, Set
-from uuid import UUID
+from uuid import uuid4
 
 from argon2.exceptions import VerifyMismatchError
 from pydantic import BaseModel
 from sqlmodel import JSON, Column, Field, Session, SQLModel, create_engine, select
 
 from lib.enums import Permission, default_permission
+from lib.env import DB_URL
 from lib.errors import (
     InvalidPassword,
+    InvalidRoleHex,
     InvalidRoleName,
     InvalidUsername,
     MissingError,
@@ -24,44 +25,59 @@ from lib.hash import hash_func, verify_func
 
 
 class DBUserBase(SQLModel):
-    __tablename__ = "user" # type: ignore
-    id: str = Field(default_factory=lambda: UUID().__str__(), primary_key=True)
+    __tablename__ = "user"  # type: ignore
+    id: str = Field(default_factory=lambda: uuid4().__str__(), primary_key=True)
     username: str = Field(unique=True, primary_key=True)
+
 
 class DBUser(DBUserBase, table=True):
     password: str
     roles: List[str] = Field(default=[], sa_column=Column(JSON))
 
+
 class User(DBUserBase):
     permissions: List[Permission] = Field(default=[], sa_column=Column(JSON))
+    roles: List[str]
+
 
 class APIUser(BaseModel):
     username: str
     password: str
 
 
+class APIUpdateUser(BaseModel):
+    username: str
+    password: str | None
+
+
 class DBRole(SQLModel, table=True):
-    __tablename__ = "role" # type: ignore
-    id: str = Field(default_factory=lambda: UUID().__str__(), primary_key=True)
+    __tablename__ = "role"  # type: ignore
+    id: str = Field(default_factory=lambda: uuid4().__str__(), primary_key=True)
     name: str = Field(unique=True)
     permissions: List[int] = Field(sa_column=Column(JSON))
+    hex: str
+
 
 class APIRole(BaseModel):
     name: str
     permissions: list[Permission]
+    hex: str
 
 
-engine = create_engine(os.getenv("DB_URL", "sqlite:///database.db"))
+engine = create_engine(DB_URL)
 SQLModel.metadata.create_all(engine)
 
 
+"""
+UTILS
+"""
+
+
 def _convert_permissions_to_enum(permission_values: List[int]) -> List[Permission]:
-    """Convert list of permission values to Permission enum objects"""
     return [Permission(value) for value in permission_values]
 
 
 def _convert_permissions_to_values(permissions: List[Permission]) -> List[int]:
-    """Convert list of Permission enum objects to values"""
     return [p.value for p in permissions]
 
 
@@ -70,20 +86,53 @@ def ensure_default():
         get_role("@admin")
     except RoleNotFound:
         _new_role(
-            DBRole(id="@admin", name="Admin", permissions=[Permission.Administrator.value])
+            DBRole(
+                id="@admin",
+                name="Admin",
+                permissions=[Permission.Administrator.value],
+                hex="#FFFFFF",
+            )
         )
 
     try:
         get_role("@everyone")
     except RoleNotFound:
         _new_role(
-            DBRole(id="@everyone", name="@everyone", permissions=[p.value for p in default_permission])
+            DBRole(
+                id="@everyone",
+                name="@everyone",
+                permissions=[p.value for p in default_permission],
+                hex="#FFFFFF",
+            )
         )
 
     try:
         get_user(id="admin")
     except UserNotFound:
-        _new_user(DBUser(id="admin", username="admin", password="ADMIN", roles=["@admin"]))
+        _new_user(
+            DBUser(id="admin", username="admin", password="ADMIN", roles=["@admin"])
+        )
+
+
+hex_regex = re.compile(r"^#[A-F0-9]{6}$")
+username_regexp = re.compile(r"^[a-z0-9_]+$")
+rolename_regexp = re.compile(r"^[a-zA-Z0-9_]+$")
+
+
+def _check_valid_username(username: str):
+    return username_regexp.match(username)
+
+
+def _check_password_safety(password: str):
+    return password.__len__() >= 8
+
+
+def _check_role_name(name: str) -> bool:
+    return bool(rolename_regexp.match(name))
+
+
+def _check_role_hex(hex: str) -> bool:
+    return bool(hex_regex.match(hex))
 
 
 """
@@ -106,12 +155,13 @@ def _calculate_permission(user: DBUser):
 
     return list(permissions)
 
+
 def get_users() -> List[User]:
     with Session(engine) as session:
         statement = select(DBUser)
         users = list(session.exec(statement).all())
         formatted_users: List[User] = []
-        
+
         for user in users:
             permissions = _calculate_permission(user)
             formatted_users.append(User(**user.model_dump(), permissions=permissions))
@@ -171,6 +221,7 @@ def verify_user(
 
 # NEW
 
+
 def _new_user(new_user: DBUser):
     with Session(engine) as session:
         try:
@@ -179,6 +230,8 @@ def _new_user(new_user: DBUser):
                 raise UserExisted()
         except UserNotFound:
             ...
+
+        new_user.password = hash_func(new_user.password)
 
         session.add(new_user)
         session.commit()
@@ -192,24 +245,13 @@ def new_user(new_user: APIUser):
 
     if not _check_password_safety(new_user.password):
         raise InvalidPassword()
-        
+
     return _new_user(DBUser(username=new_user.username, password=new_user.password))
-
-
-username_regexp = re.compile(r"^[a-z0-9_]+$")
-
-
-def _check_valid_username(username: str):
-    return username_regexp.match(username)
-
-
-def _check_password_safety(password: str):
-    return password.__len__() >= 8
 
 
 # UPDATE
 def update_user(
-    new_user: APIUser, id: str | None = None, username: str | None = None
+    new_user: APIUpdateUser, id: str | None = None, username: str | None = None
 ):
     with Session(engine) as session:
         db_user = get_user_db(id=id, username=username, session=session)
@@ -220,16 +262,32 @@ def update_user(
                 continue
 
             if db_user_dump[key] != value:
-                if key == "username" and not _check_valid_username(value):
-                    raise InvalidUsername()
+                if key == "username":
+                    if not _check_valid_username(value):
+                        raise InvalidUsername()
 
-                if key == "password":
-                    if not _check_password_safety(new_user.password):
+                    try:
+                        existed_user = get_user(username=value)
+                        if existed_user.id != db_user.id:
+                            raise UserExisted()
+
+                    except UserNotFound:
+                        ...
+
+                if key == "password" and value:
+                    if not _check_password_safety(value):
                         raise InvalidPassword()
 
                     value = hash_func(value)
 
-                setattr(db_user, key, value)
+                if value:
+                    setattr(db_user, key, value)
+
+        if not _check_valid_username(db_user.username):
+            raise InvalidUsername()
+
+        if not _check_password_safety(db_user.password):
+            raise InvalidPassword()
 
         session.add(db_user)
         session.commit()
@@ -256,19 +314,20 @@ def user_has_permission(user: User, permissions: List[Permission]) -> bool:
     everyone = get_role(id="@everyone")
 
     for permission in permissions:
-        if permission in everyone.permissions:
+        if permission.value in everyone.permissions or permission in user.permissions:
             continue
 
-        permission_bit = permission.value.__int__()
+        permission_bit = permission.value
 
         if permission_bit >= 10:
-            s_prefix, s_suffix = permission_bit.__str__().split("", 1)
+            permission_bit_str = permission_bit.__str__()
+            s_prefix, s_suffix = permission_bit_str[0], permission_bit_str[1:]
             prefix = int(s_prefix)
             suffix = int(s_suffix)
 
-            if (
-                suffix != 0
-                and Permission(prefix * s_suffix.__len__()) in user.permissions
+            if suffix != 0 and (
+                Permission(prefix * (10 ** s_suffix.__len__())) in user.permissions
+                or prefix * (10 ** s_suffix.__len__()) in everyone.permissions
             ):
                 continue
 
@@ -322,6 +381,7 @@ def get_role(
 
 # NEW
 
+
 def _new_role(new_role: DBRole):
     with Session(engine) as session:
         try:
@@ -338,17 +398,20 @@ def _new_role(new_role: DBRole):
 
 
 def new_role(new_role: APIRole):
-    if not check_role_name(new_role.name):
+    if not _check_role_name(new_role.name):
         raise InvalidRoleName()
 
+    if not _check_role_hex(new_role.hex):
+        raise InvalidRoleHex()
+
     return _new_role(
-        DBRole(name=new_role.name, permissions=_convert_permissions_to_values(new_role.permissions))
+        DBRole(
+            name=new_role.name,
+            permissions=_convert_permissions_to_values(new_role.permissions),
+            hex=new_role.hex,
+        )
     )
 
-rolename_regexp = re.compile(r"^[a-zA-Z0-9_]+$")
-
-def check_role_name(name: str) -> bool:
-    return bool(rolename_regexp.match(name))
 
 # UPDATE
 def update_role(new_role: APIRole, id: str | None = None, name: str | None = None):
@@ -356,7 +419,9 @@ def update_role(new_role: APIRole, id: str | None = None, name: str | None = Non
         db_role = get_role(id=id, name=name, session=session)
 
         if db_role.id == "@admin":
-            raise NotAllowed()
+            _new_role = APIRole(**db_role.model_dump())
+            _new_role.hex = new_role.hex
+            new_role = _new_role
 
         db_role_dump = db_role.model_dump()
 
@@ -364,19 +429,32 @@ def update_role(new_role: APIRole, id: str | None = None, name: str | None = Non
             if key == "id":
                 continue
 
+            if key == "permissions":
+                value = [permission.value for permission in value]
+
             if db_role_dump[key] != value:
                 setattr(db_role, key, value)
 
         session.add(db_role)
         session.commit()
 
-        return get_role(id=id, name=new_role.name, session=session)
+    if not _check_role_name(db_role.name):
+        raise InvalidRoleName()
+
+    if not _check_role_hex(db_role.hex):
+        raise InvalidRoleHex()
+
+    return get_role(id=id, name=new_role.name, session=session)
 
 
 # DELETE
 def delete_role(id: str | None = None, name: str | None = None):
     with Session(engine) as session:
         exist_role = get_role(id=id, name=name, session=session)
+
+        if exist_role.id == "@admin" or exist_role.id == "@everyone":
+            raise NotAllowed("you cant delete @admin or @everyone role")
+
         session.delete(exist_role)
         session.commit()
 
@@ -384,10 +462,11 @@ def delete_role(id: str | None = None, name: str | None = None):
 # UTILITY
 def update_roles(user: User, roles: List[DBRole]):
     with Session(engine) as session:
-        db_user = DBUser(**user.model_dump())
-        
+        db_user = get_user_db(id=user.id, session=session)
+
         db_user.roles = [role.id for role in roles]
 
         session.add(db_user)
+        session.commit()
 
         return get_user_db(id=user.id, session=session)
